+++ src/lib/primevue/components/menubar/sizes/index.ts
@@ -1,7 +1,7 @@
 const sizes = {
-  gap: '0.5rem',
-  padding: '0.5rem 0.75rem',
-  borderRadius: '{content.border.radius}',
+  gap: '1em',
+  padding: '0.5rem 1rem',
+  borderRadius: '0rem',
 };
 
 export default sizes