+++ src/lib/primevue/semantic/color-scheme/text/light/index.ts
@@ -3,6 +3,7 @@ export default {
     color: '{surface.750}',
     hoverColor: '{surface.800}',
     mutedColor: '{surface.500}',
-    hoverMutedColor: '{surface.600}'
+    hoverMutedColor: '{surface.600}',
+    avatarColor: '{gray.0}'
   }
 }